trace-tests: Fix finding of debugfs directory
The debugfs directory search used the first field of the /proc/mounts
file, but that's not the actual name of the type of mount, but the name
used in the mount command. Use the type to test, as if the mount name
changes, it will fail to find the debugfs directory.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>