ftrace-test-saved_cmdlines: Fix to handle new logic
The saved_cmdline logic in the kernel was extended and can now
handle a lot more cmdlines. Fix the test to not expect an exact amount
of saved cmdlines as the kernel now rounds up the number it can save.
Just make sure that it saves at least what is expected.

Also stop checking for <...>, where the saved cmdlines didn't save
something. The new logic is much less likely to not save something.

Shorten the sleep time from 5 to 1 second, as we could sleep for a
minute and still not find <...> missing command lines.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>